pyproject.toml
[project]
name = "duku-recs"
version = "0.1.0"
description = "Local MF/ALS training + FastAPI serving"
requires-python = ">=3.9"
dependencies = [
  "numpy>=1.26",
  "pandas>=2.0",
  "scipy>=1.11",
  "implicit>=0.7.2",
  "fastapi>=0.110",
  "uvicorn[standard]>=0.29",
  "pydantic>=2.5",
  "pyyaml>=6.0",
  "joblib>=1.4"
]

[tool.uvicorn]
app = "serve.app:app"
host = "127.0.0.1"
port = 8000
reload = true
